+++ packages/cli/src/lib/uploadMessages.ts
@@ -6,7 +6,20 @@ import { AppMessages } from '@appsemble/types';
 import axios from 'axios';
 import yaml from 'js-yaml';
 
-export async function uploadMessages(path: string, appId: string, remote: string): Promise<void> {
+/**
+ * Upload messages for an app.
+ *
+ * @param path - The path to the app directory.
+ * @param appId - The app id to upload the messages for.
+ * @param remote - The remote to upload the messages to.
+ * @param force - Whether or not to force the update for locked apps.
+ */
+export async function uploadMessages(
+  path: string,
+  appId: string,
+  remote: string,
+  force: boolean,
+): Promise<void> {
   if (!existsSync(join(path, 'messages'))) {
     return;
   }
@@ -25,7 +38,7 @@ export async function uploadMessages(path: string, appId: string, remote: string
     const language = parse(messageFile).name;
     const file = await fs.readFile(join(path, 'messages', messageFile), 'utf8');
     const messages = yaml.safeLoad(file);
-    result.push({ language, messages } as AppMessages);
+    result.push({ force, language, messages } as AppMessages);
   }
 
   for (const language of result) {