+++ packages/server/src/migrations/index.ts
@@ -28,6 +28,7 @@ import * as m29 from './0.17.3';
 import * as m30 from './0.17.6';
 import * as m31 from './0.17.7';
 import * as m32 from './0.18.0';
+import * as m33 from './0.18.1';
 import * as m1 from './0.9.0';
 import * as m2 from './0.9.3';
 import * as m3 from './0.9.4';
@@ -65,4 +66,5 @@ export const migrations: Migration[] = [
   m30,
   m31,
   m32,
+  m33,
 ];