+++ packages/studio/src/pages/blocks/block/index.tsx
@@ -14,7 +14,6 @@ import { defaultLocale } from '@appsemble/utils';
 import { Fragment, ReactElement, useCallback } from 'react';
 import { FormattedMessage, useIntl } from 'react-intl';
 import { Redirect, useHistory, useRouteMatch } from 'react-router-dom';
-import { Definition } from 'typescript-json-schema';
 
 import { ActionTable } from './ActionTable';
 import { EventTable } from './EventTable';
@@ -147,23 +146,21 @@ export function BlockPage(): ReactElement {
       {selectedBlockManifest.parameters?.definitions && (
         <>
           <Title level={4}>{untranslatedMessages.definitions}</Title>
-          {Object.entries((selectedBlockManifest.parameters as any).definitions).map(
-            ([key, definition]: [string, Definition]) => (
-              <Fragment key={key}>
-                <Title lang={defaultLocale} level={5}>
-                  <a href={`${url}#${key}`} id={key}>
-                    {key}
-                  </a>
-                </Title>
-                {definition.description && <MarkdownContent content={definition.description} />}
-                {definition.type === 'object' || definition.type === 'array' ? (
-                  <ParameterTable parameters={definition} />
-                ) : (
-                  <TypeTable definition={definition} />
-                )}
-              </Fragment>
-            ),
-          )}
+          {Object.entries(selectedBlockManifest.parameters.definitions).map(([key, definition]) => (
+            <Fragment key={key}>
+              <Title lang={defaultLocale} level={5}>
+                <a href={`${url}#${key}`} id={key}>
+                  {key}
+                </a>
+              </Title>
+              {definition.description && <MarkdownContent content={definition.description} />}
+              {definition.type === 'object' || definition.type === 'array' ? (
+                <ParameterTable parameters={definition} />
+              ) : (
+                <TypeTable definition={definition} />
+              )}
+            </Fragment>
+          ))}
         </>
       )}
     </Content>